separate the total comments and reservations file from totalmeals.js
