+++ src/modules/totalReservations.js
@@ -1,0 +1,11 @@
+import { getReservations } from "./reservations.js";
+
+export const getTotalReservations = async (itemId) => {
+  const reservations = await getReservations(itemId);
+  let totalReservations = 0;
+
+  reservations.forEach(() => {
+    totalReservations++;
+  });
+  return totalReservations;
+};
